confirm message product

+++ cleardb.php
@@ -9,8 +9,10 @@
  */
 	$query2 =  "Delete from $dbname.Sales where SalesID > 0"; 	
 	$result = mysqli_query($conn,$query2);	
-	
-
+	Product_Grouping
+	querychange = "ALTER TABLE Sales MODIFY Product_Grouping DECIMAL";
+	$result_querychange = mysqli_query($conn,$querychange);	
+	if(!result_querychange){echo "ERROR";}
 	
 		$query3 =  "Delete from $dbname.Product_Component where ComponentID > 0"; 	
 		$result3 = mysqli_query($conn,$query3);
